angular.app/stark/packages/stark-core/src/validation/decorators/is-bic/is-bic.validator.decorator.spec.ts
/*tslint:disable:completed-docs*/
import { validateSync, ValidationError } from "class-validator";
import { StarkIsBIC } from "./is-bic.validator.decorator";
import { starkIsBICValidatorName } from "../../validators/is-bic";

class MyClass {
	@StarkIsBIC()
	public name: string;
}

class SimpleClass {
	@StarkIsBIC()
	public dummyObject: object;
}

describe("ValidatorDecorator: StarkIsBIC", () => {
	let myClass: MyClass;
	let simpleClass: SimpleClass;
	const validatorConstraintName: string = starkIsBICValidatorName;

	beforeEach(() => {
		myClass = new MyClass();
		simpleClass = new SimpleClass();
	});

	it("should fail if the object to validate is not a string", () => {
		const errors: ValidationError[] = validateSync(simpleClass);

		expect(errors.length).toBe(1);
		expect(errors[0].constraints).toBeDefined();
		expect(errors[0].constraints[validatorConstraintName]).toBeDefined();
	});

	it("should fail if bic is empty", () => {
		const errors: ValidationError[] = validateSync(myClass);

		expect(errors.length).toBe(1);
		expect(errors[0].constraints).toBeDefined();
		expect(errors[0].constraints[validatorConstraintName]).toBeDefined();
	});

	it("should fail if bic is not correct", () => {
		myClass.name = "123EBEBBTRE";
		const errors: ValidationError[] = validateSync(myClass);

		expect(errors.length).toBe(1);
	});

	it("should NOT fail if bic is correct", () => {
		myClass.name = "NBBEBEBBTRE";
		const errors: ValidationError[] = validateSync(myClass);

		expect(errors.length).toBe(0);
	});
});
